Inject validator/selector_builder/finder as locator dependencies
This allow to override the default behavior and isolate nested classes
