+++ lib/watir-webdriver/locators/finders/text_area_finder.rb
@@ -1,0 +1,11 @@
+module Watir
+  class TextAreaLocator
+    class Finder < ElementLocator::Finder
+      private
+
+      def can_convert_regexp_to_contains?
+        false
+      end
+    end
+  end
+end
