+++ lib/watir-webdriver/locators/selector_builders/child_cell_selector_builder.rb
@@ -1,0 +1,22 @@
+module Watir
+  class ChildCellLocator
+    class SelectorBuilder < ElementLocator::SelectorBuilder
+      def build_wd_selector(selectors)
+        return if selectors.values.any? { |e| e.kind_of? Regexp }
+
+        expressions = %w[./th ./td]
+        attr_expr = attribute_expression(selectors)
+
+        unless attr_expr.empty?
+          expressions.map! { |e| "#{e}[#{attr_expr}]" }
+        end
+
+        xpath = expressions.join(" | ")
+
+        p build_wd_selector: xpath if $DEBUG
+
+        [:xpath, xpath]
+      end
+    end
+  end
+end
